requirements.txt
Flask>=3.0
Flask-Cors>=4.0
bcrypt>=4.0
pyjwt>=2.8
pydantic>=2.7
pydantic-settings>=2.0
python-dotenv>=1.0
supabase>=0.21
psycopg2-binary>=2.9
sqlalchemy>=2.0
